CB-20455 Separate application and quartz db connection pool. With the separation of the db connection pool used by quartz the application itself we can minimize the impact on the application caused by If we separate the connection pool used by quartz and the connection pool used by the application itself, then we can minimize the impact caused by misbehaving quartz.
In order to keep the implementation simple we will go with the same default for the new connection pool what we have for application pool. This means that in certain conditions the maximum connections towards the database can be doubled, which is acceptable.
