+++ core/src/main/java/com/sequenceiq/cloudbreak/conf/DatabaseConfig.java
@@ -7,13 +7,11 @@ import java.sql.SQLException;
 
 import javax.inject.Inject;
 import javax.inject.Named;
-import javax.sql.DataSource;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
-import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.Environment;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.JpaVendorAdapter;
@@ -26,6 +24,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import com.sequenceiq.cloudbreak.common.database.BatchProperties;
 import com.sequenceiq.cloudbreak.common.database.JpaPropertiesFacory;
 import com.sequenceiq.cloudbreak.common.tx.CircuitBreakerType;
+import com.sequenceiq.cloudbreak.database.MultiDataSourceConfig;
 import com.sequenceiq.cloudbreak.util.DatabaseUtil;
 import com.sequenceiq.flow.ha.NodeConfig;
 import com.zaxxer.hikari.HikariConfig;
@@ -33,7 +32,7 @@ import com.zaxxer.hikari.HikariDataSource;
 
 @Configuration
 @EnableTransactionManagement
-public class DatabaseConfig {
+public class DatabaseConfig extends MultiDataSourceConfig {
 
     @Value("${cb.db.env.user:}")
     private String dbUser;
@@ -87,23 +86,7 @@ public class DatabaseConfig {
     @Inject
     private Environment environment;
 
-    @Bean(name = "defaultDataSource")
-    public DataSource defaultDataSource() throws SQLException {
-        return getDataSource("hikari-app-pool");
-    }
-
-    @Bean(name = "quartzDataSource")
-    public DataSource quartzDataSource() throws SQLException {
-        return getDataSource("hikari-quartz-pool");
-    }
-
-    @Primary
-    @Bean(name = "dataSource")
-    public DataSource dataSource() {
-        return new RoutingDataSource();
-    }
-
-    private HikariDataSource getDataSource(String poolName) throws SQLException {
+    protected HikariDataSource getDataSource(String poolName) throws SQLException {
         DatabaseUtil.createSchemaIfNeeded("postgresql", databaseAddress, dbName, dbUser, dbPassword, dbSchemaName);
         HikariConfig config = new HikariConfig();
         config.setPoolName(poolName);
