+++ core/src/test/java/com/sequenceiq/cloudbreak/database/RoutingDataSourceTest.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.conf;
+package com.sequenceiq.cloudbreak.database;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.lenient;
